requirements.txt
# ─────────────────────────────────────────────
# SanjeevaniRxAI Backend – Production Requirements
# ─────────────────────────────────────────────

# ── Web Framework ─────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0

# ── Database ──────────────────────────────────
pymongo[srv]==4.7.2
motor==3.4.0                  # async MongoDB driver (for future async endpoints)

# ── Data Processing ───────────────────────────
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.4               # Excel (.xlsx) read/write support

# ── Validation & Settings ─────────────────────
pydantic==2.7.4
pydantic-settings==2.3.0
python-dotenv==1.0.1

# ── ML / AI ───────────────────────────────────
scikit-learn==1.5.0
joblib==1.4.2

# ── LLM Integration ───────────────────────────
httpx==0.27.0                 # async HTTP client (Groq SDK + test client)
requests==2.32.3              # sync HTTP client for outbound notifications
twilio==9.3.6                 # WhatsApp/SMS TwiML response support

# ── Rate Limiting ─────────────────────────────
slowapi==0.1.9
limits==3.13.0

# ── Caching ───────────────────────────────────
cachetools==5.3.3

# ── Resilience & Retry ────────────────────────
tenacity==8.3.0

# ── Authentication ────────────────────────────
PyJWT==2.8.0                  # JWT token generation/validation
google-auth==2.29.0           # Google OAuth verification
google-auth-oauthlib==1.2.0   # Google OAuth flow
google-auth-httplib2==0.2.0   # Google auth HTTP transport

# ── Serialization / Utilities ─────────────────
python-multipart==0.0.9       # FastAPI form/file upload
email-validator==2.1.1        # Pydantic email fields
pytz==2024.1                  # Timezone handling

# ── Testing (install locally only; not needed on Render) ──────────────────
# pytest==8.2.2
# pytest-asyncio==0.23.7
# pytest-cov==5.0.0

# ── Development / Linting (optional but useful) ─
# black==24.4.2
# ruff==0.4.7

# ── Agentic Framework (Hackathon Requirement) ───
langchain>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0
langchain-groq>=0.2.0
langchain-anthropic>=0.3.0
langfuse>=2.50.0


groq==0.9.0
httpx==0.27.0                 # async HTTP client (Groq SDK + test client)
requests==2.32.3              # sync HTTP client for outbound notifications
twilio==9.3.6
telegram